LIBSEARCH-63. Prepended "RDF" module to "register_alias" definitions
For consistency, modified the "register_alias" method in classes in the
"types" to include the "RDF" module.

This consistency is necessary because the classes in the "types"
subdirectory may be loaded in an arbitrary order. Previous to this
change, the loading of the classes could fail with a
"NameError: uninitialized constant" if a class without the "RDF"
module specified (such as "int.rb") was loaded first.

Prepending the "RDF" module to all register_alias calls where it is
needed prevents this error from occurring, no matter the order in which
the classes are loaded.

https://issues.umd.edu/browse/LIBSEARCH-63